dev-requirements.txt
pytest-ckan
pytest-cov
six>=1.12.0